Updates to models
