+++ SuiteCareers/SuiteCareers/Models/Question.cs
@@ -1,0 +1,11 @@
+using System.ComponentModel.DataAnnotations;
+
+public class Question
+{
+    [Key]
+    public int questionId { get; set; }
+
+    [Required]
+    public string? question { get; set; }
+}
+
