+++ SuiteCareers/SuiteCareers/Models/Response.cs
@@ -1,0 +1,10 @@
+using System.ComponentModel.DataAnnotations;
+
+public class Response
+{
+    [Key]
+    public int responseId { get; set; }
+    public int response { get; set; }
+
+    public int questionId { get; set; }
+}
