+++ SuiteCareers/SuiteCareers/Models/SweetCareerDbContext.cs
@@ -1,0 +1,2 @@
+
+public class SweetCareerDbContext : DbContext
