+++ SuiteCareers/SuiteCareers/Models/User.cs
@@ -3,8 +3,7 @@ using System.ComponentModel.DataAnnotations.Schema;
 
 namespace SuiteCareers.Models
 {
-    public class UserInfo
-    {
+    
         public class User
         {
 
@@ -20,68 +19,15 @@ namespace SuiteCareers.Models
             public string? State { get; set; }
         }
 
-        public class UserDescription
-        {
-            [Key]
-            public int DescriptionId {get; set;}
-            [Required(ErrorMessage = "Please enter your educational level")]
-            public string EducationLevel { get; set;}
-            [Required(ErrorMessage = "Please enter your work experience")]
-            public string WorkExperience { get; set; }
-            [Required(ErrorMessage = "Please enter your job description")]
-            public string UserJob { get; set;}
-            [Required(ErrorMessage = "Please enter the date")]
-            public DateTime date { get; set; }
-            //How to add foreigh key to this class?
-            //[DataType(DataType.EmailAddress)]
-            //public string? Email { get; set; }
+     
 
-        }
-        public class Session
-        {
-            [Key]
-            public int sessionId { get; set; }
-            [Required(ErrorMessage = "Please enter the date")]
-            public DateTime date { get; set; }
-            [Required]
-            public int interviewId { get; set; }
+        
 
-            public string email { get; set; }
-
+        
             
-        }
-
-        public class Interview
-        {
-            [Key]
-            public int interviewId { get; set; }
-
-
-            [Required]
-            public int questionId { get; set; }
-
-
-        }
-
-        public class Question
-        {
-            [Key]
-            public int questionId { get; set; }
-
-            [Required]
-            public string? question { get; set; }
-        }
-            
-        public class Response
-        {
-            [Key]
-            public int responseId { get; set; }
-            public int response { get; set; }
-            
-            public int questionId { get; set; }
-        }
+        
 
 
 
     }
-}
+
